third_party/blink/web_tests/external/wpt/css/css-typed-om/the-stylepropertymap/computed/iterable.tentative-expected.txt

This is a testharness.js-based test.
PASS StylePropertyMap iterates properties in correct order
PASS StylePropertyMap iterator returns CSS properties with the correct CSSStyleValue
PASS StylePropertyMap iterator returns list-valued properties with the correct CSSStyleValue
FAIL StylePropertyMap iterator returns custom properties with the correct CSSStyleValue assert_equals: expected "A" but got " A"
PASS Computed StylePropertyMap sorts custom properties in increasing code-point order
Harness: the test ran to completion.
